pyproject.toml
[build-system]
requires      = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name         = "reverb-gui"
version      = "0.1.0"
description  = "GUI front-end for Rev’s Reverb ASR+diarization pipelines"
readme       = "README.md"
license      = "Apache-2.0"
authors      = ["Your Name <you@example.com>"]
keywords     = ["asr","diarization","reverb","gui"]
classifiers  = [
  "Development Status :: 4 - Beta",
  "Environment :: X11 Applications :: Qt",
  "Programming Language :: Python :: 3.10",
  "Topic :: Multimedia :: Sound/Audio :: Speech",
]
packages = [
  { include = "wenet", from = "asr" },
  { include = "reverb_gui" },
]

[tool.poetry.dependencies]
python            = ">=3.10,<3.14"
torch             = {version = "2.2.2+cu121", source = "pytorch_cuda"}
torchvision       = {version = "0.17.2+cu121", source = "pytorch_cuda"}
torchaudio        = {version = "2.2.2+cu121", source = "pytorch_cuda"}
openai-whisper    = "*"
typeguard         = "^2.0"
wandb             = "*"
pyyaml            = ">=3.12"
huggingface-hub   = "^0.22.2"
numpy             = "<2"
GitPython         = "*"
python-dotenv     = "^1.0.1"
"pyannote.audio"  = "==3.3.1"
intervaltree      = "==3.1.0"
PySide6           = "^6.7"
ffmpeg-python     = "^0.2"
pyannote-database = "^5.1.3"
tqdm              = "^4.66.2"

[tool.poetry.group.dev.dependencies]
pytest      = "^8.1.1"
ruff        = "^0.1"
pytest-qt   = "^4.4.0"
pyinstaller = "^6.6.0"

[tool.poetry.scripts]
reverb     = "wenet.bin.recognize_wav:main"
reverb-gui = "reverb_gui.main:launch_gui"

[[tool.poetry.source]]
name     = "pytorch_cuda"
url      = "https://download.pytorch.org/whl/cu121"
priority = "explicit"
